pyproject.toml
[project]
name = "incident-commander"
version = "0.1.0"
description = "Autonomous Incident Commander with Multi-Agent Architecture"
requires-python = ">=3.12"
dependencies = [
    "langgraph==1.0.7",
    "langchain==1.2.6",
    "langchain-openai>=0.2.0",
    "langfuse>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
